feat: Bump eslint-plugin-react (#52)
We have an eslint-config package that depends on the @callstack/eslint-config, and adds a few overrides and eslint-plugin-react-redux. 
We had a need to update eslint-plugin-react to 7.13.0, which then requires  eslint-plugin-import  as a peerDependency. This means the apps have to add the import plugin to their package.json.

Bumping the package version here so the peerDependency is encapsulated here.